add email validation

+++ server/src/services/user/DeleteUserService.ts
@@ -1,10 +1,18 @@
 import prisma from "../../database";
+import validateEmail from "../../util/emailValidate";
 
 class DeleteUserService{
     async handle(email: string){
         if (!email){
             throw new Error("Please, fill in all fields");
         }
+
+        const isEmailValid = validateEmail(email);
+
+        if (!isEmailValid){
+            throw new Error("Insert a valid email");
+        }
+
         const deletePosts = prisma.task.deleteMany({
             where: {
               authorEmail: email,
